moved function calculateRevisionDate to module common.mjs  . added test for this function
